[Fix]: Global sockets
Changed following
	-implemented socket event listeners of chat and video class globally
