+++ client/src/components/student/chatComponents/Chat.js
@@ -3,14 +3,14 @@ import NavBar from '../mainComponents/NavBar'
 import {Container,Row,Col} from 'react-bootstrap'
 import ChatBox from './ChatBox'
 import Sidebar from '../mainComponents/Sidebar'
-export default function Chat() {
+export default function Chat(props) {
     return (
     <React.Fragment>
         <NavBar/>
         <Container fluid>
         <Row >
             <Col sm={8}className="d-flex mh100 justify-content-center" > 
-                <ChatBox/>     
+                <ChatBox {...props}/>     
             </Col>
             <Col sm={4}className="d-flex justify-content-center" >
                 <Sidebar/>
